@bep 2015-05-13
@spf13 - what's your thoughts on this one? It is an important use case. One could argue that this should be done at render time, but there is a wormhole of challenges (mainly due to the render once, use twice) with that approach. And the performance penalty from this one is the same as for `canonifyUrls`.