@spf13 2015-05-15
I think it's a nice addition and there's a solid use case for making the site portable. Since the performance impact isn't much I'd say this is solid. LGTM.